Fixed a path error.
The ISO/IEC 14443A component is at components/iso_iec_14443A (uppercase
A), but the makefile used iso_iec_14443a (lowercase a). The scp and
rsync tools seemed happy with this but the resulting folder on the
server had a lowercase a.

When however you check out the repository directly on the server the
folder had the uppercase A, and the synthesis script didn't work since
it looked for files in iso_iec_14443a/ which isn't present.