LIBPREP: Don't rename the .gds PG ports, rename the .lef ones instead.
In the libprep script we read the library in 3 different formats. The
.gds which has PG ports vdd! and gnd!, the LEF which has PG ports vdd
and gnd, and the DB which also has PG ports vdd and gnd.

Previously I was reading the .gds first, renaming the PG ports to vdd
and gnd, and then reading the other files. This worked nicely until I
got to looking at LVS.

The LVS process involves taking the netlist and merging it with .cdl
netlist to get a transister layer netlist. However the CDL also uses
vdd! and gnd!, and I couldn't find a way to get the tools to link the
vdd/gnd ports in my netlist with the vdd!/gnd! ports of the CDL.

So instead I'm renaming the PG ports in the LEF to match the GDS. We
have to do this manually so that the -merge_action attributes_only
argument works as desired and can update the attributes on the PG pins
correctly.

The tools auto rename the PG ports in the DBs to match the physical
library.