DP: Updating power plan.
I found that three of the std_cell rails were not connected to power/gnd.
This was due to the opposite (vdd for gnd and gnd for vdd) stripe being
on top of that rail and so there was no way to connect it up, or at
least the tools didn't do it automatically.

To remedy this I extended the std_cell rails to the outermost ring,
which fixed two of the rails, and generally seemed like a good idea.
Then I separated the stripes a bit more, which allowed enough space to
connect in that last rail.

Finally I added VDD and VSS terminals to my design.