DP: Added pin constraints.
preferred_port_locations.tcl is the output of write_pin_constraints
which is run after place_pins -self.

On subsequent runs of this script this file can be read to attempt to
keep pins in the same locations. If that file doesn't exist then the
manual constraints are run.

So far the manual constraints are relatively loose. I constrain the
connections to the analog block to be on the top edge. The connections
to the sensor to be on the right hand edge in the top half, and the
connections to the ADC in the bottom half of the right hand edge. With
the uid_variable[2:0] signals being anywhere as long as they are
together.

We could tighten / change these constraints later once we get a better
idea on what the other blocks will look like.