PNR: Output the .gds both with and without the layermap.
XFAB provides and expects GDSII files to use a different layer numbering
to what the synopsys tools use.

I'm outputting both formats now. The XFAB format is needed for the LVS
script. The ICC2 format could be used to import directly into the larger
design to connect it to the other blocks.