DP: Reworked the floorplan.
I was having a bunch of issues with metal density DRC errors, and
noticed that the output of DRC with: -D DUMMY_FILL -D DUMMY_FILL_RESULTS
included a bunch of co-ordinates that were out of the boundary of my
floorplan (260x260).

It turns out that the rings don't go in the core area, and so we need to
use core_offset to leave space for the rings.

The core_utilization argument is only used when combined with the
side_ratio argument, instead of explicitly stating the side lengths, it
creates the core sized to meet the desired core_utilization rate.

I also adjusted the pin constraints, and the PG mesh to relate to the
new floorplan size.