DP/PNR: On creating a new block, save the old block first.
This means you can load the block from a particular section and get the
state at the end of that section.

In PNR in the ICV In-Design section we have to manually save the block
several times. So without this change loading the pnr_icv block would
load the state half way through the section. This change makes the
behaviour consistant over all sections.