SYNTH: Changed how the set_false_paths work.
I had initially set up these paths to cut paths going into the
synchronisers. This means there would be warnings about unconstrained
paths if someone were to use an async input directly, rather than the
synchronised version. This works great in synthesis, but unfortunately
when the constraints get translated to the ICC2 format for use in the
DP/PNR scripts they got converted to "-to [get_ports ...]", the DP
script then gave an error when reading these constraints that you can't
use -to with an input port.

I couldn't find a work around for that, so instead I've added a helper
function which creates the false path and outputs a list of cut paths.
This allows us to see in the logs that we aren't accidentally cutting
too many paths. Which is better than nothing.