VERIFICATION: Adding an archive of coverage metrics.
Last year I generated coverage metrics for all my modules from all my
tests. I didn't bother to check it in, because I could always
re-generate it.

I'm currently unable to run simulations with VCS. This appears to be due
to a licensing issue. Therefore I'm checking this archive in now. This
old data is valid, at least until I modify any of the code.

This is mostly as a backup, so the data will be available when I come to
writing up the thesis.