RTL: Changed how the IDENTIFY reply works.
Previously I had a couple of parameters passed into the top level design
from the synthesis script, that specify the versions of the sensor and
the ADC.

I realised that this approach would mean that changing just the ADC or
th sensor would require re-running synthesis, and pnr of my block, which
isn't ideal. Once my block is finished we want to sign off on the GDSII
and not need to change it, unless internal edits are required.

So instead I've made these input ports. The other blocks can hard wire
them to indicate their versions.

Additionally I reduced them to 4 bits, because if we get to more than 16
versions of a design, we have bigger issues.

Finally I split the iso_iec_14443a_version into two parts, the digital
iso14443a IP core and the analogue fronte end AFE part.