RTL: Minor RTL changes in order to generate a complete .saif.
Also updating the ISO/IEC 14443A component to pull a commit that does
the same as described below.

To do basic power analisation and optimisation, it is necessary to have
some info about the switching activity of the design. We can get this
information by generating a .saif file from simulation.

VCS was missing signals in interfaces and enum signals when they were
assigned to directly from a sequential block.

The solution in the case of enums is to declare the variables as type
logic [X:0], we can still use the enum types to assign to it.

In the case of interfaces, I created temporary signals which I assign to
in the sequential block, and then connect them to the interface signals
with wires.