Updating the ISO/IEC 14443A core to latest.
This is a change to fix some testbenches.

When I changed the initialisation module's state signal from an enum to logic [1:0] it broke state name reporting in the tests, since state no longer has the .name property.

To fix this I moved the State enum into the ISO14443A_pkg so I can access that type from the testbenches and cast the state signal back to that type.